Started Day 8 of AoC 2022
Got as far as reading each line and checking for the maximum value in each line. Unsure how to iterate through to find the furthest occurrence of the maximum value in each line, and then count characters that come after. I think I could use the 'index' variable to record position and use a boolean operand to test if the maximum value occurs more than once, but outside of keeping a variable going that records the number of items (which seems lazy) I don't know what to do next.